+++ pyphs/tests/cppTest.py
@@ -12,7 +12,7 @@ from __future__ import division
 # retrieve the pyphs.PHSCore of a nonlinear RLC from the tutorial on PHSCore
 from pyphs.tutorials.phscore import core as nlcore
 
-from pyphs import PHSNumericalMethod, PHSNumericalCore, numcore2cpp
+from pyphs import PHSNumericalCore, numcore2cpp
 import os
 import shutil
 
@@ -20,12 +20,12 @@ here = os.path.realpath(__file__)[:os.path.realpath(__file__).rfind(os.sep)]
 
 
 def cpp_nlcore_full():
-    method = PHSNumericalMethod(nlcore)
-    nums = PHSNumericalCore(method)
+    nums = PHSNumericalCore(nlcore)
 
     label = 'test'
     path = os.path.join(here, label)
 
     numcore2cpp(nums, objlabel=label, path=path)
-    shutil.rmtree(path)
+    if not os.name.lower().startswith('nt'):
+        shutil.rmtree(path)
     return True
